Update the Kbuild file to fix header includes
To include the soc/gs headers we are using EXTRA_CFLAGS in the Makefile.
One caveat with this is that ccflags-y is checked first before
EXTRA_CFLAGS. For this project, we have some header stubs under
include/linux which are added in Kbuild using ccflags-y. This means that
the stubs are actually getting used instead of the real headers from
private/google-modules/soc/gs/include. So fix the Kbuild to remove the
local includes for all but the trace header in order to properly include
the real kernel headers.

This fixes PM stability and firmware loading issues on this branch. Here
are the logs observed:

  I edgetpu_platform 1ce00000.abrolhos: Initial power state: 2
  E edgetpu_platform 1ce00000.abrolhos: pm_runtime_get_sync returned 1
  I edgetpu abrolhos: Powering up
  E edgetpu abrolhos: edgetpu_ioctl_finalize_group: pm_get failed (1)

Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I7984d61f38932ac27f9ce2a8bb31481f3c27e3a8