Support downloading prebuilt GKI fips140.ko
Add build options and flags to toggle downloading GKI fips140.ko
which is enabled by default in this change.

Currently there's a limitation that we can only download the file
and place it in output directory; however, the `modules.load` doesn't
include this module. Therefore, the `fips140.ko` won't be loaded
at boot time with the images generated in kernel source tree.

In order to load the module at boot time, it should be placed in
AOSP kernel prebuilt directory and re-pack the images.

Bug: 331791671
Change-Id: I9a7a8dedb4cfc4b4492d45db5d50d7fc468c41dc
Signed-off-by: Qian-Hao Huang <qhhuang@google.com>