build.config.gs101: scheduler vendor hooks are available on v6.1 now
androidboot.vh_sched_available=0 is meant to inform the Android HAL
that scheduler vendor hooks are not available, which was the case when
this was introduced originally as part of b/265961701

Since then, they've been introduced on v6.1, so the flag here should be
dropped.

Otherwise locally built Pixel modules will cause the HAL to not benefit
from these hooks.

Test: TH
Bug: 352501631
Change-Id: I913d9be8990d7e287fc2bd5d421f8daa9e40b4bc
Signed-off-by: André Draszik <draszik@google.com>