pcie: exynos: Remove pcie phy reset on resume
There is an issue on whi/pro that it may get stuck while resetting PCIe
PHY on resume. On zuma the step for resetting PCIe PHY on resume was
removed. Consulted with PHY team and confirmed that this change can be
applied to whi/pro as well.

Bug: 360724998
Change-Id: I3a599b44ff882dc86a88b612d6354b59247d9302
Merged-In: I3a599b44ff882dc86a88b612d6354b59247d9302
Signed-off-by: tinghsin <tinghsin@google.com>