google_modemctl: New modem interface module
Exynos modems expose a few notifiers and hooks for client modules. These
interfaces are the same across modems, but they currently can't coexist.
Pull them out into a control module, so we don't require per-modem-model
dependencies between various Google modules and <model>-cpif. The
infrastructure is almost there (a shared
<soc/google/exynos-modem-ctrl.h>), but we flesh it out with some of the
common implementation. Divest the "exynos" name while we're at it,
because these operations are fairly generic at this point.

Most of this code is carried from existing cpif modules, but there are
some small tweaks along the way:
 - removing unused headers
 - using RAW_NOTIFIER_HEAD()
 - turning modem_force_crash_exit_ext() into a notifier, to allow
   multiple implementations

Note that modem_notify_event() carries a stub / TODO; it's partially
wired up in the modem drivers, but it's not actually used to effect any
real action for listeners. I choose not to unwind this ball of wax for
now, and simply carry the APIs over for compatibility.

Bug: 360934165
Test: build for various s5300, s5400 devices; modem still lights up

Change-Id: Id3c467dde0b40d059e07877ffef2de25ee34e31a
Signed-off-by: Brian Norris <briannorris@google.com>