alsa: Drop direct cpif dependencies
These dependencies are artificial. The only substantial header that is
supposed to be exported here is modem_notifier.h, which we don't use.
All our modem dependencies are brokered through
<soc/google/modem_notifier.h>, which is via gs_soc_headers, not
cpif_headers.

Also include a new optional dependency [1] for GOOGLE_MODEMCTL while
we're at it, since some APIs are getting shuffled around to a new module
in google-modules/soc/gs.

[1] https://docs.kernel.org/kbuild/kconfig-language.html#optional-dependencies

Bug: 360934165
Test: build for various SoCs
Test: `modinfo -F depends aoc_alsa_dev_util.ko` contains
      exynos_modemctl, not cpif

Change-Id: Id6eff5f825972cbee2eb5dd890ca20bbebac9554
Signed-off-by: Brian Norris <briannorris@google.com>