drm: samsung: Adjust DFP/sink count check for DisplayPort
During testing with third-party USB hubs, we found that there are cases
where the number of downstream-facing-ports (DFP) is 0, but the number
of sinks is reported as 1, which currently causes DisplayPort link to
fail with the "invalid sink count" error. Relaxing this check allows
those hubs to support DisplayPort without issues.

Bug: 346778638
Test: tested DisplayPort with third-party USB hubs
Signed-off-by: Sergey Volk <servolk@google.com>
Change-Id: I2209d73f842fc295f429ffee770137cd412bc664