touch/usi: fix setting the validity of battery level
Set the validity of the battery level whenever there's a
battery level reporting.

Currently, with 0 as initial value, we set the validity only
when the level has changed. But due to this, after switching stylus,
if the new stylus has the same battery level with the previous one,
there's error in reading the level until the level gets changed.

Bug: 374163591

Change-Id: Ib4fdc7d493030a0b7816ab55a259af501b57e188
Signed-off-by: Hyungwoo Yang <hyungwooyang@google.com>