lwis_util: Use sched_setscheduler_nocheck()
The "checked" version is for user-space requests. This is kernel code.
All we're doing is setting ourselves up for needless SELinux denials,
depending on the context of what process (e.g., init, vendor_init,
ueventd) is doing the module-load, or whether device probe is deferred
to a worker thread.

See also b/131252752#comment12 for similar recommendations, which led to
https://git.kernel.org/linus/04e048cf09d7b5fc995817cdc5ae1acd4482429c

Side note: these APIs are mostly deprecated, especially for modules, and
are only available as an Android-ism to keep them exported. We should
probably move to
sched_set_fifo()/sched_set_fifo_low()/sched_set_normal(), which *are*
still available for modules, and already use the 'nocheck' variant
internally.

Bug: 360934165

Change-Id: Iaca46777777b47e906c1c38fda3a11b6abbc46ba
Signed-off-by: Brian Norris <briannorris@google.com>