vh: sched: Fix definition of check_cred()
We only care about whether the task has CAP_SYS_NICE to allow modifying
the attributes. The current code is suitable for verifying cgroupv1
migration where it was copied from. But for scheduler usage,
CAP_SYS_NICE is sufficient check.

Modify all users as we no longer require to pass any arg because we just
need to check if current task which is doing the modification has
a privilege to do so.

Bug: 353988746
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: If0a11f21dc7dc07a683e3027b631ed35b14c49d8