vh: sched: Always protect raw_spin_locks with _irqsave variant [ DO NOT MERGE ANYWHERE ]
We can't rely on the caller to have disabled the interrupts. Many of
scheduler operations can happen from interrupt context e.g:
load_balance() or an IPI. Instead of relying on carefully auditing the
code to know we need and irqsave variant or not, enforce it by default.

Since flags name clashes with arg name in some function, use irqflags
everywhere to pass to the _irqsave functions.

Bug: 382033293
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I4aa5018495ee5e8ab98682d0377da8f5ce19414a